o implements: "when clicking and holding down the mouse button on the magnyfiers, a list with a set of available thumbnail sizes pops up." from the roadmap (if that was intendet to be so).
I think it doesn't matter, that the 10% evaluates to 20% actually due to the min and max size settings from Constans. The enduser is most probably intrested in a rough sizing idea, rather then having exact measures, anayway.

Currently the size selections are static and the same on "+" and "-". One could extand that behaviour, to dynamically show only larger sizes then current on '+' and smaller ones then current on '-'.