o adds the comment as ToolTip to the TagTree.
If we wish to have a tooltip for the sourcetree, too, we should build a "class Tree : KListView" and derive {Source,Tag}Tree from that. Additionally we should have a "class TreeNode : KListViewItem", from which we derive our {Source,Tag}Treenodes then. This is just for clean object oriented code design.