o cleanup o fix up of the overlays o overlays are now switchable on/off via contextMenu o drawing of the overlays is now part of the ImageViewer Widget, and not the XImages as it was before. o some changes in memory footprint
though I think, that the ImageViewer is pretty usable already, this is not finished yet. The code needs to be cleaned, and translation stuff and configuration stuff needs to be added. I'll start looking into that, as soon as  the exif stuff in the tree is integrated, as I like to add some exif infos to the info overlay :-).