Fairly huge commit again....
o added a secrecy function, so one can set a tag to be 'secret' (dialogmanagetag) and when the filter is either set to include or to ignore, one is asked, if those images really should be displayed. No, I don't want my friends to see some images again while testing ;-)
o the above point brought a fairly big shuffeling and redesign in the tagtreenode stuff
	o moved the filter to TagTreeNode and with it most functions that every subclass did define itself.
o additionally neccessery xml additions were added to save the secrecy status of the tags.
o added the imageimporter menuitem, though the importer DOES NOT WORK YET. this needs a reinstall of kphotobook to install the new kphotobookui.rc file.