- icon for exif tag added - new class DateTimeFilterData holds all infos about a datefilter - all dates are stored in ISO format now - menus and actions are now longer instantiated in KPhotobook class but in ActionProvider and MenuProvider - the currently set filter is now stored even if the data is not stored
Unfortunately I did not succeed in my attempt to use a XML based contextmenu for the imageviewer. I lost a lot of time in trying different dirty hacks to make it work. I beleive I should use an own GUIFactory/GUIClient for the imageviewer. But I'm not very motivated for doing that...